Adds unhandled event hook to state machine
Allows state machines to define an optional hook
(OnUnhandledEvent) that is called when an event bubbles
all the way to the root without being handled.

This provides a mechanism for observing, logging, or
tracing unhandled events. Includes a test case to verify
that the hook is only called for truly unhandled events.